requirements.txt
pdfplumber==0.10.4
python-docx==1.1.2
PyYAML==6.0.2
lxml==5.3.1
streamlit==1.35.0
protobuf==3.20.3
pymupdf>=1.23.0
pillow>=10.0.0
